update to the routes, handlebars, and seeds
