+++ controllers/api/user-routes.js
@@ -36,7 +36,6 @@ router.get('/:id', (req, res) => {
 
 router.post('/', (req, res) => {
   User.create({
-    username: req.body.username,
     email: req.body.email,
     password: req.body.password
   })
@@ -44,7 +43,6 @@ router.post('/', (req, res) => {
   .then(dbUserData => {
     req.session.save(() => {
       req.session.user_id = dbUserData.id;
-      req.session.username = dbUserData.username;
       req.session.loggedIn = true;
 
       res.json(dbUserData);
@@ -77,7 +75,6 @@ router.post('/login', (req, res) => {
   
       req.session.save(() => {
         req.session.user_id = dbUserData.id;
-        req.session.username = dbUserData.username;
         req.session.loggedIn = true;
   
         res.json({ user: dbUserData, message: 'logged in!' });
