+++ controllers/home-routes.js
@@ -10,4 +10,8 @@ router.get('/login', (req, res) => {
     res.render('login');
 });
 
+router.get('/createaccount', (req, res) => {
+    res.render('createaccount');
+});
+
 module.exports = router;
