+++ index.html
@@ -20,33 +20,31 @@
         </nav>
     </header>
 
-    <main>
-        <main class="container">        
-            <div class="row center">           
-                <div id="party-carousel" class="col carousel ">
-                    <div class="row carousel-item" href="#one">
-                        <div class="col s12">
-                            <img class="img" src="./public/images/DnD3.jpg">
-                            <span class="black"><a href="./party.html">Ocean Adventures</a></span>
-                        </div>
+    <main class="container">        
+        <div class="row center">           
+            <div id="party-carousel" class="col carousel ">
+                <div class="row carousel-item" href="#one">
+                    <div class="col s12">
+                        <img class="img" src="./public/images/DnD3.jpg">
+                        <span class="black"><a href="./party.html">Ocean Adventures</a></span>
                     </div>
-    
-                    <div class="row carousel-item" href="#two">
-                        <div class="col s12">
-                            <img class="img" src="./public/images/DnD.jpg">
-                            <span class="black"><a href="./party.html">Dragon Hunters</a></span>
-                        </div>
+                </div>
+
+                <div class="row carousel-item" href="#two">
+                    <div class="col s12">
+                        <img class="img" src="./public/images/DnD.jpg">
+                        <span class="black"><a href="./party.html">Dragon Hunters</a></span>
                     </div>
-    
-                    <div class="row carousel-item" href="#three">
-                        <div class="col s12">
-                            <img class="img" src="./public/images/DnD2.jpg">
-                            <span class="black"><a href="./party.html">Glory and Honor</a></span>
-                        </div>
+                </div>
+
+                <div class="row carousel-item" href="#three">
+                    <div class="col s12">
+                        <img class="img" src="./public/images/DnD2.jpg">
+                        <span class="black"><a href="./party.html">Glory and Honor</a></span>
                     </div>
                 </div>
             </div>
-        </main>
+        </div>
     </main>
 
     <footer>
