+++ public/javascript/index.js
@@ -3,3 +3,4 @@ $(document).ready(function () {
 });
 
 
+document.querySelector('.signup-form').addEventListener('submit', signupFormHandler);
