+++ public/javascript/login.js
@@ -1,11 +1,11 @@
 async function loginFormHandler(event) {
     event.preventDefault();
   
-    const email = document.getElementById('#email').value.trim();        
-    const password = document.getElementById('#password').value.trim();  
+    const email = document.getElementById('email').value.trim();        
+    const password = document.getElementById('password').value.trim();  
   
     if (email && password) {
-      const response = await fetch('/login', {
+      const response = await fetch('/api/users/login', {
         method: 'post',
         body: JSON.stringify({
           email,
@@ -22,31 +22,7 @@ async function loginFormHandler(event) {
     }
   }
   
-  async function signupFormHandler(event) {
-      event.preventDefault();
+
     
-      const email = document.getElementById('#email').value.trim();
-      const password = document.getElementById('#password').value.trim();
+    document.querySelector('#login-form').addEventListener('submit', loginFormHandler);
     
-      if (username && email && password) {
-        const response = await fetch('/users', {
-          method: 'post',
-          body: JSON.stringify({
-            username,
-            email,
-            password
-          }),
-          headers: { 'Content-Type': 'application/json' }
-        });
-    
-        if (response.ok) {
-          document.location.replace('/');
-        } else {
-          alert(response.statusText);
-        }
-      }
-    }
-    
-    document.querySelector('.login-form').addEventListener('submit', loginFormHandler);
-    
-    document.querySelector('.signup-form').addEventListener('submit', signupFormHandler);
