+++ server.js
@@ -21,9 +21,8 @@ const sess = {
 
 app.use(session(sess));  
 
-const helpers = require('./utils/helpers');
 
-const hbs = exphbs.create({ helpers });
+const hbs = exphbs.create({});
 
 app.engine('handlebars', hbs.engine);
 app.set('view engine', 'handlebars');
@@ -39,7 +38,7 @@ const routes = require('./controllers');
 
 //handlebars GET
 app.get('/', (req, res) => {
-  res.render('home');
+  res.render('homepage');
 })
 
 sequelize.sync({ force: false }).then(() => {
