+++ views/createaccount.handlebars
@@ -20,7 +20,7 @@
             <div class='row'>
               <div class='input-field col s12'>
                 <input class='validate orange-text text-accent-2' type='email' name='email' id='email' />
-                <label for='email'>Enter your username</label>
+                <label for='email'>Enter your email</label>
               </div>
             </div>
 
@@ -44,4 +44,6 @@
 
     <div class="section"></div>
     <div class="section"></div>
+
+    <script src="/javascript/create-user.js"></script>
   </main>
