+++ views/login.handlebars
@@ -14,7 +14,7 @@
 
             <div id="form" class="z-depth-1 lighten-4 row" style="display: inline-block; padding: 32px 48px 0px 48px;">
 
-                <form class="col s12" method="post">
+                <form id="login-form" class="col s12" method="post">
 
                     <div class='row'>
                         <div class='col s12'>
@@ -24,7 +24,7 @@
                     <div class='row'>
                         <div class='input-field col s12'>
                             <input class='validate orange-text text-accent-2' type='email' name='email' id='email' />
-                            <label for='email'>Enter your username</label>
+                            <label for='email'>Enter your email</label>
                         </div>
                     </div>
 
@@ -54,5 +54,5 @@
 </main>
 
 
-<script src="/js/login.js"></script>
-<script src="/js/logout.js"></script>
+<script src="/javascript/login.js"></script>
+<script src="/javascript/logout.js"></script>
